Backported from MIPS (Cobalt).

+++ drivers/isdn/avmb1/b1capi.c
@@ -6,6 +6,9 @@
  * (c) Copyright 1997 by Carsten Paeth (calle@calle.in-berlin.de)
  * 
  * $Log$
+ * Revision 1.4.2.18  1998/03/20 20:34:37  calle
+ * port valid check now only for T1, because of the PCI and PCMCIA cards.
+ *
  * Revision 1.4.2.17  1998/03/20 14:38:17  calle
  * capidrv: prepared state machines for suspend/resume/hold
  * capidrv: fix bug in state machine if B1/T1 is out of nccis
@@ -595,7 +598,7 @@ int avmb1_registercard(int port, int irq, int cardtype, int allocio)
 				 SA_SHIRQ, card->name, card)) != 0) {
 		printk(KERN_ERR "b1capi: unable to get IRQ %d (irqval=%d).\n",
 		       irq, irqval);
-		release_region((unsigned short) port, AVMB1_PORTLEN);
+		release_region(port, AVMB1_PORTLEN);
 		return -EBUSY;
 	}
 
@@ -650,7 +653,7 @@ int avmb1_detectcard(int port, int irq, int cardtype)
 
 int avmb1_probecard(int port, int irq, int cardtype)
 {
-	if (check_region((unsigned short) port, AVMB1_PORTLEN)) {
+	if (check_region(port, AVMB1_PORTLEN)) {
 		printk(KERN_WARNING
 		       "b1capi: ports 0x%03x-0x%03x in use.\n",
 		       port, port + AVMB1_PORTLEN);
@@ -1219,7 +1222,6 @@ int avmb1_init(void)
 {
 	char *p;
 	char rev[10];
-
 
 #ifndef HAS_NEW_SYMTAB
 	/* No symbols to export, hide all symbols */
